% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscales.r
\docType{data}
\name{pprabhu_palettes}
\alias{pprabhu_palettes}
\title{Customized color palettes for presentations}
\format{An object of class \code{list} of length 7.}
\usage{
pprabhu_palettes
}
\description{
A collection of colour palettes for data science projects for presentations.
main : basic color palette
cool : color palette with cool colors
hot : color palette with warmer colors
div : color palette that puts equal emphasis on mid-range critical values and extremes at both ends of the data range.
sseq : single sequence palette that are suited to ordered data that progress from low to high.
mseq : multiple sequence palette that are suited to ordered data that progress from low to high.
qual : color palette that creates the primary visual differences between classes.
}
\keyword{datasets}
