% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscales.r
\name{scale_fill_pprabhu}
\alias{scale_fill_pprabhu}
\title{Scales for ggplot2 : scale_fill function}
\usage{
scale_fill_pprabhu(..., palette = "main", discrete = TRUE,
  reverse = TRUE)
}
\arguments{
\item{...}{additional arguments to pass to scale_fill_gradientn}

\item{palette}{Choose from 'pprabhu_palettes' list}

\item{discrete}{whether to use a discrete/continuous colour palette}

\item{reverse}{logical input, Do you want to reverse the order of colors?}
}
\description{
Applies your custom color scales as fill aesthetics
}
\examples{
library(ggplot2)
# Create some variables
x <- 1:9 * 2
y <- 1:9 * 1
xy <- data.frame(x,y)

# Plots a histogram with discrete fill using "div" palette
ggplot(xy,aes(x=x,y=y,fill=as.character(y))) +
  geom_bar(stat="identity") +
  scale_fill_pprabhu(palette = "div")

# Plots a histogram with continious fill using "hot" palette
ggplot(xy, aes(x = x)) +
     geom_histogram(aes(fill = ..count..), binwidth = 5) +
     scale_fill_pprabhu(palette = "hot",discrete=FALSE)

}
